feat: #14 컨트롤러에서 토큰 받기
