+++ src/main/java/com/example/creditmarket/controller/MyPageController.java
@@ -1,9 +1,10 @@
-package com.example.creditmarket.controller.mypage;
+package com.example.creditmarket.controller;
 
 import com.example.creditmarket.dto.FavoriteResponseDTO;
 import com.example.creditmarket.dto.OrderResponseDTO;
 import com.example.creditmarket.service.mypage.MyPageService;
 import lombok.RequiredArgsConstructor;
+import org.springframework.security.core.Authentication;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PatchMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -17,19 +18,18 @@ public class MyPageController {
 
     private final MyPageService myPageService;
 
-    //토큰 받기전에 임시로 email 받기
     @GetMapping("/favor/{page}")
-    public List<FavoriteResponseDTO> selectFavoriteList(@PathVariable int page, String userEmail) {
-        return myPageService.selectFavoriteList(page, userEmail);
+    public List<FavoriteResponseDTO> selectFavoriteList(@PathVariable int page, Authentication authentication) {
+        return myPageService.selectFavoriteList(page, authentication);
     }
 
     @GetMapping("/buy/{page}")
-    public List<OrderResponseDTO> selectOrderList(@PathVariable int page, String userEmail) {
-        return myPageService.selectOrderList(page, userEmail);
+    public List<OrderResponseDTO> selectOrderList(@PathVariable int page, Authentication authentication) {
+        return myPageService.selectOrderList(page, authentication);
     }
 
     @PatchMapping("/buy/{orderId}")
-    public String updateOrder(@PathVariable Long orderId, String userEmail) {
-        return myPageService.updateOrder(orderId, userEmail);
+    public String updateOrder(@PathVariable Long orderId, Authentication authentication) {
+        return myPageService.updateOrder(orderId, authentication);
     }
 }
